Fix error reporting

+++ src/com/confusinguser/sbgods/remoteguildchat/RemoteGuildChat.java
@@ -20,9 +20,9 @@ public class RemoteGuildChat {
     private String latestGuildmessage = "";
     private final List<String> blockedMessages = new ArrayList<>();
 
-    public void handleGuildMessage(String author, String message, HypixelRank rank, InetAddress requestSenderIpAddr, boolean showMessageOnly, MessageChannel channel) {
+    public void handleGuildMessage(String author, String message, HypixelRank rank, InetAddress requestSenderIpAddr, boolean showMessageOnly, boolean guildPrefix, MessageChannel channel) {
         if (SBGods.getInstance().getDiscord() == null || author.isEmpty() || message.isEmpty()) return;
-        boolean shouldSendMessage = !blockedMessages.contains(author + ":" + message) && !latestGuildmessage.equals("Guild > " + author + ": " + message) &&
+        boolean shouldSendMessage = !blockedMessages.contains(author + ":" + message) && !latestGuildmessage.equals((guildPrefix ? "Guild > " : "") + author + ": " + message) &&
                 !latestGuildmessageAuthor.equals(author) &&
                 !latestMessageByIP.getOrDefault(requestSenderIpAddr, "").equals(author + ":" + message);
 
@@ -31,7 +31,7 @@ public class RemoteGuildChat {
             latestMessageByIP.put(requestSenderIpAddr, author + ":" + message);
             blockedMessages.add(SBGods.getInstance().getUtil().stripColorCodes(author + ":" + message));
             Multithreading.scheduleOnce(() -> blockedMessages.remove(author + ":" + message), 12, TimeUnit.SECONDS);
-            sendMessage(author, message, rank, showMessageOnly, channel);
+            sendMessage(author, message, rank, showMessageOnly, guildPrefix, channel);
         }
     }
 
@@ -49,11 +49,10 @@ public class RemoteGuildChat {
         }
     }
 
-    private void sendMessage(String author, String message, HypixelRank rank, boolean showMessageOnly, MessageChannel channel) {
+    private void sendMessage(String author, String message, HypixelRank rank, boolean showMessageOnly, boolean guildPrefix, MessageChannel channel) {
         if (channel != null) {
-            if (latestGuildmessageAuthor.equals(author)) {
-                if (showMessageOnly) latestGuildmessage += "\nGuild > " + message;
-                else latestGuildmessage += "\nGuild > " + author + ": " + message;
+            if (latestGuildmessageAuthor.equals(author) && !showMessageOnly) {
+                latestGuildmessage += "\nGuild > " + author + ": " + message;
                 try {
                     channel.deleteMessageById(channel.getLatestMessageId()).queue();
                 } catch (IllegalStateException ignored) {
